Several changes--genericness of code etc.
I made several changes to make the code more generic and also to make
the engine crate slightly easier to work with.

1.) I changed what gets returned from the initialize_animal_list to an
enclosure.  We treat this as the temporary enclosure (id 0) so that we
can treat all movement of animals as having a source and a target
enclosure.

2.) I had to restructure the testing to comply with the structural
changes I made.

3.) I changed the way that I deal with the animal and enclosure structs.
Rather than having them in separate subdirs I moved them into the src
dir for the crate and named them accordingly.  I think this is slightly
easier to follow.